rfct: move context menu into its own space
